+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changes
 
 ### Fixes
+- Fixed orientation not getting reset after crashing while flying and getting back on the Pokémon.
 - Fixed Furfrou not being trimmable on NeoForge.
 
 ### Developer